+++ src/HealthCheckBundle.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Lgarret\HealthCheckBundle;
 
 use Lgarret\HealthCheckBundle\DependencyInjection\Compiler\BuiltinChecksCompilerPass;
+use Lgarret\HealthCheckBundle\DependencyInjection\Compiler\HealthCheckClientCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -15,5 +16,6 @@ final class HealthCheckBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new BuiltinChecksCompilerPass());
+        $container->addCompilerPass(new HealthCheckClientCompilerPass());
     }
 }